backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
langchain==0.0.352
langchain-openai==0.0.2
pillow==10.1.0
numpy==1.26.2
langgraph==0.0.15
torch==2.2.2
torchvision==0.17.2
transformers==4.35.2
opencv-python==4.8.1.78
python-jose[cryptography]
passlib[bcrypt] 